+++ tests/mapped_struct.py
@@ -1584,6 +1584,19 @@ class MappedFrozensetPackingTest(unittest.TestCase, CommonCollectionPackingTest)
         b = buffer("m")
         self.assertRaises(IndexError, self.PACKING_CLASS.unpack_from, b, 0)
 
+    def testBitmapSets(self):
+        a = bytearray(16)
+        bitmap_values = [
+            frozenset([1,3,6]),
+            frozenset([10,30,60]),
+            frozenset([69,99]),
+            frozenset([1,5,7,38,49,67,99,105,119]),
+            frozenset([119]),
+        ]
+        for fs in bitmap_values:
+            mapped_struct.mapped_frozenset.pack_into(fs, a, 0)
+            self.assertEqual(mapped_struct.mapped_frozenset.unpack_from(a, 0), fs)
+
 class MappedListPackingTest(unittest.TestCase, CommonCollectionPackingTest, IndexedCollectionPackingTest):
     PACKING_CLASS = mapped_struct.mapped_list
     COLLECTION_CLASS = list